+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 3.28.0
+
+- Add support for creating and updating PayPal customer session
+- Add support for getting PayPal customer recommendations
+
 ## 3.27.0
 
 - Add recipient/contact info: `recipient_email`and `recipient_phone` to `Transaction